Repeat CryptoStream.Read() for .NET 6 compat
.NET 6 changed `System.Security.Cryptography.CryptoStream.Read()` such
that it may produce short reads (i.e. read less than the requested
number of bytes) before the end of the stream is reached.  This is
documented in [Partial and zero-byte reads in DeflateStream, GZipStream,
and CryptoStream][partial-reads].

To support .NET 6 and later versions, call .Read() repeatedly, until the
expected number of bytes has been read or end-of-stream has been
reached.

[partial-reads]: https://learn.microsoft.com/dotnet/core/compatibility/core-libraries/6.0/partial-byte-reads-in-streams

Fixes: #23
Signed-off-by: Kevin Locke <kevin@kevinlocke.name>
Based-on: https://github.com/hadeutscher/OvercookedTool/pull/24
Based-on-patch-by: Sergej Salnikov <sergejkiller@gmail.com>